+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 Change Log
 ==========
 
+### 0.3.0 (2017-07-25)
+
 - 新增 Configuration\GetEnv 類別
 - 新增 DateTime\Clock 類別及 3 個具體時鐘類別：SystemClock、FixedClock 與 OffsetClock
 - 新增 DateTime\Duration 類別
+- Slim\Controller::checkParams() 增加檢查 PARAM_ARRAY
 
 ### 0.2.1 (2017-07-10)